Move the lr_cactus function to the parser.
This will be needed by other error recovery mechanisms.
