+++ src/lib/corchuelo.rs
@@ -35,7 +35,7 @@ use std::convert::{TryFrom, TryInto};
 use std::fmt::Debug;
 
 use cactus::Cactus;
-use cfgrammar::{NTIdx, Symbol, TIdx};
+use cfgrammar::Symbol;
 use lrlex::Lexeme;
 use lrtable::{Action, StIdx};
 
@@ -110,8 +110,8 @@ pub(crate) fn recover<TokId: Clone + Copy + Debug + TryFrom<usize> + TryInto<usi
                                                                         .unwrap(),
                                                          next_lexeme.start(), 0);
                             let (new_la_idx, n_pstack) =
-                                lr_cactus(parser, Some(new_lexeme), la_idx, la_idx + 1,
-                                          pstack.clone(), None);
+                                parser.lr_cactus(Some(new_lexeme), la_idx, la_idx + 1,
+                                                 pstack.clone(), &mut None);
                             if new_la_idx > la_idx {
                                 debug_assert_eq!(new_la_idx, la_idx + 1);
                                 let n_repairs = repairs.child(ParseRepair::Insert{term_idx});
@@ -161,7 +161,7 @@ pub(crate) fn recover<TokId: Clone + Copy + Debug + TryFrom<usize> + TryInto<usi
         //   \wedge (j = N \vee (0 <= j < N \wedge f(q_r, t_{j + 1} \in {accept, error}))
         {
             let (new_la_idx, n_pstack)
-                = lr_cactus(parser, None, la_idx, la_idx + PARSE_AT_LEAST, pstack.clone(), None);
+                = parser.lr_cactus(None, la_idx, la_idx + PARSE_AT_LEAST, pstack.clone(), &mut None);
             if new_la_idx < in_la_idx + PORTION_THRESHOLD {
                 // A repair is a "finisher" (i.e. can be considered complete and doesn't need to be
                 // added to the todo list) if it's parsed at least N symbols or parsing ends in
@@ -228,15 +228,15 @@ pub(crate) fn recover<TokId: Clone + Copy + Debug + TryFrom<usize> + TryInto<usi
                                                                 .ok()
                                                                 .unwrap(),
                                                  next_lexeme.start(), 0);
-                    pstack = lr_cactus(parser, Some(new_lexeme), la_idx, la_idx + 1,
-                                       pstack, Some(tstack)).1;
+                    pstack = parser.lr_cactus(Some(new_lexeme), la_idx, la_idx + 1,
+                                              pstack, &mut Some(tstack)).1;
                 },
                 ParseRepair::Delete => {
                     la_idx += 1;
                 }
                 ParseRepair::Shift => {
                     let (new_la_idx, n_pstack)
-                        = lr_cactus(parser, None, la_idx, la_idx + 1, pstack, Some(tstack));
+                        = parser.lr_cactus(None, la_idx, la_idx + 1, pstack, &mut Some(tstack));
                     assert_eq!(new_la_idx, la_idx + 1);
                     la_idx = new_la_idx;
                     pstack = n_pstack;
@@ -254,61 +254,6 @@ pub(crate) fn recover<TokId: Clone + Copy + Debug + TryFrom<usize> + TryInto<usi
     }
 
     (la_idx, repairs)
-}
-
-/// Start parsing text at `la_idx` (using the lexeme in `lexeme_prefix`, if it is not `None`,
-/// as the first lexeme) up to (but excluding) `end_la_idx`. If an error is encountered, parsing
-/// immediately terminates (without recovery).
-///
-/// Note that if `lexeme_prefix` is specified, `la_idx` will still be incremented, and thus
-/// `end_la_idx` *must* be set to `la_idx + 1` in order that the parser doesn't skip the real
-/// lexeme at position `la_idx`.
-fn lr_cactus<TokId: Clone + Copy + Debug + TryFrom<usize> + TryInto<usize> + PartialEq>
-    (parser: &Parser<TokId>, lexeme_prefix: Option<Lexeme<TokId>>, mut la_idx: usize, end_la_idx: usize,
-     mut pstack: Cactus<StIdx>, mut tstack: Option<&mut Vec<Node<TokId>>>)
-  -> (usize, Cactus<StIdx>)
-{
-    assert!(lexeme_prefix.is_none() || end_la_idx == la_idx + 1);
-    while la_idx != end_la_idx {
-        let st = *pstack.val().unwrap();
-        let (la_lexeme, la_term) = parser.next_lexeme(lexeme_prefix, la_idx);
-
-        match parser.stable.action(st, la_term) {
-            Some(Action::Reduce(prod_id)) => {
-                let nonterm_idx = parser.grm.prod_to_nonterm(prod_id);
-                let pop_num = parser.grm.prod(prod_id).unwrap().len();
-                if let Some(ref mut tstack_uw) = tstack {
-                    let nodes = tstack_uw.drain(pstack.len() - pop_num - 1..)
-                                         .collect::<Vec<Node<TokId>>>();
-                    tstack_uw.push(Node::Nonterm{nonterm_idx: nonterm_idx, nodes: nodes});
-                }
-
-                for _ in 0..pop_num {
-                    pstack = pstack.parent().unwrap();
-                }
-                let prior = *pstack.val().unwrap();
-                pstack = pstack.child(parser.stable.goto(prior, NTIdx::from(nonterm_idx)).unwrap());
-            },
-            Some(Action::Shift(state_id)) => {
-                if let Some(ref mut tstack_uw) = tstack {
-                    tstack_uw.push(Node::Term{lexeme: la_lexeme});
-                }
-                pstack = pstack.child(state_id);
-                la_idx += 1;
-            },
-            Some(Action::Accept) => {
-                debug_assert_eq!(la_term, Symbol::Term(TIdx::from(parser.grm.eof_term_idx())));
-                if let Some(ref mut tstack_uw) = tstack {
-                    debug_assert_eq!(tstack_uw.len(), 1);
-                }
-                break;
-            },
-            None => {
-                break;
-            }
-        }
-    }
-    (la_idx, pstack)
 }
 
 fn score(repairs: &Cactus<ParseRepair>) -> usize {
